Updated schema.sql
Changed price fields type to decimal

+++ core/src/main/resources/db/schema.sql
@@ -16,7 +16,7 @@ create table phones (
   id BIGINT AUTO_INCREMENT primary key,
   brand VARCHAR(50) NOT NULL,
   model VARCHAR(254) NOT NULL,
-  price FLOAT,
+  price DECIMAL(10,2),
   displaySizeInches FLOAT,
   weightGr SMALLINT,
   lengthMm FLOAT,
@@ -66,9 +66,9 @@ create table statuses(
 
 create table orders (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    subtotal FLOAT NOT NULL,
-    deliveryPrice FLOAT NOT NULL,
-    totalPrice DECIMAL(6,2) NOT NULL,
+    subtotal DECIMAL(10,2) NOT NULL,
+    deliveryPrice DECIMAL(10,2) NOT NULL,
+    totalPrice DECIMAL(10,2) NOT NULL,
     firstName VARCHAR(20) NOT NULL,
     lastName VARCHAR(25) NOT NULL,
     deliveryAddress VARCHAR(50) NOT NULL,
@@ -81,7 +81,7 @@ create table phone2order(
     phoneId BIGINT,
     orderId BIGINT,
     quantity BIGINT,
-    price DECIMAL (6,2),
+    price DECIMAL(10,2),
     CONSTRAINT FK_phone2order_phoneId FOREIGN KEY (phoneId) REFERENCES phones (id) ON DELETE CASCADE ON UPDATE CASCADE,
     CONSTRAINT FK_phone2order_orderId FOREIGN KEY (orderId) REFERENCES orders (id) ON DELETE CASCADE ON UPDATE CASCADE
 );
